requirements.txt
asgiref==3.5.2
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
distlib==0.3.6
Django==4.1.4
django-jalali-date==1.0.2
django-modeltranslation==0.18.10
django-summernote==0.8.20.0
djangorestframework==3.14.0
filelock==3.8.2
idna==3.4
jdatetime==4.1.1
mysqlclient==2.1.1
Pillow==9.4.0
platformdirs==2.6.0
PyMySQL==1.0.2
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.13
virtualenv==20.17.1
webencodings==0.5.1
